+++ pkg/alertmanager/multitenant_test.go
@@ -82,6 +82,7 @@ func mockAlertmanagerConfig(t *testing.T) *MultitenantAlertmanagerConfig {
 	cfg.ShardingRing.InstanceID = "test"
 	cfg.ShardingRing.InstanceAddr = "127.0.0.1"
 	cfg.PollInterval = time.Minute
+	cfg.ShardingRing.FinalSleep = 0
 
 	return cfg
 }